Strictly follow language spec and remove all language extensions (#52)
Due to issues with testing and disabling language extensions, we decided to remove all of them and just strictly follow the required specification. There for this PR

    Removes support for Unicode identifiers and input files
    Removes support for line comments starting with //
    Split integer literals with leading 0 into two tokens.
    Quick fix new BigInteger integration in parser. Requires proper fix!

Closes #29
Closes #45

+++ src/main/java/com/github/firmwehr/gentle/parser/tokens/IntegerLiteralToken.java
@@ -2,12 +2,15 @@ package com.github.firmwehr.gentle.parser.tokens;
 
 import com.github.firmwehr.gentle.source.SourceSpan;
 
+import java.math.BigInteger;
+
 public record IntegerLiteralToken(
 	SourceSpan sourceSpan,
-	int value
+	BigInteger value
 ) implements Token {
 	@Override
 	public String format() {
+		// this is NOT debug output, we actually have to rely on the toString output representation
 		return "integer literal " + value;
 	}
 }
